requirements.txt
boto3
Nuitka
ordered-set
python-dotenv
PySide6
psutil
pywin32
requests